[ui-storageBrowser]Implement multiselect for Storage browser table data
