+++ desktop/core/src/desktop/js/apps/storageBrowser/StorageBrowserPage/StorageBrowserTabContents/StorageBrowserTabContent.tsx
@@ -15,27 +15,14 @@
 // limitations under the License.
 
 import React, { useState, useEffect } from 'react';
-import { Spin, Dropdown, Input } from 'antd';
-import type { MenuProps } from 'antd';
-import Button from 'cuix/dist/components/Button';
-import { PrimaryButton } from 'cuix/dist/components/Button';
+import { Spin } from 'antd';
 
 import { i18nReact } from '../../../../utils/i18nReact';
 import BucketIcon from '@cloudera/cuix-core/icons/react/BucketIcon';
-import CopyClipboardIcon from '@cloudera/cuix-core/icons/react/CopyClipboardIcon';
-import DataMovementIcon from '@cloudera/cuix-core/icons/react/DataMovementIcon';
-import DeleteIcon from '@cloudera/cuix-core/icons/react/DeleteIcon';
-import DownloadIcon from '@cloudera/cuix-core/icons/react/DownloadIcon';
-import DropDownIcon from '@cloudera/cuix-core/icons/react/DropdownIcon';
-import FolderIcon from '@cloudera/cuix-core/icons/react/ProjectIcon';
-import ImportIcon from '@cloudera/cuix-core/icons/react/ImportIcon';
-//TODO: Use cuix icon (Currently fileIcon does not exist in cuix)
-import { FileOutlined } from '@ant-design/icons';
 
 import PathBrowser from '../../../../reactComponents/FileChooser/PathBrowser/PathBrowser';
-import InputModal from '../../InputModal/InputModal';
 import StorageBrowserTable from '../StorageBrowserTable/StorageBrowserTable';
-import { fetchFiles, mkdir, touch } from '../../../../reactComponents/FileChooser/api';
+import { fetchFiles } from '../../../../reactComponents/FileChooser/api';
 import {
   PathAndFileData,
   StorageBrowserTableData,
@@ -69,102 +56,10 @@ const StorageBrowserTabContent = ({
   const [sortOrder, setSortOrder] = useState<SortOrder>(SortOrder.NONE);
   //TODO: Add filter functionality
   const [filterData] = useState<string>('');
-  const [showNewFolderModal, setShowNewFolderModal] = useState<boolean>(false);
-  const [showNewFileModal, setShowNewFileModal] = useState<boolean>(false);
+
   const [refreshKey, setRefreshKey] = useState<number>(0);
 
   const { t } = i18nReact.useTranslation();
-
-  const newActionsMenuItems: MenuProps['items'] = [
-    {
-      key: 'create',
-      type: 'group',
-      label: t('CREATE'),
-      children: [
-        {
-          icon: <FileOutlined />,
-          key: 'new_file',
-          label: t('New File'),
-          onClick: () => {
-            setShowNewFileModal(true);
-          }
-        },
-        {
-          icon: <FolderIcon />,
-          key: 'new_folder',
-          label: t('New Folder'),
-          onClick: () => {
-            setShowNewFolderModal(true);
-          }
-        }
-      ]
-    },
-    {
-      key: 'upload',
-      type: 'group',
-      label: t('UPLOAD'),
-      children: [
-        {
-          icon: <ImportIcon />,
-          key: 'upload',
-          label: t('New Upload')
-        }
-      ]
-    }
-  ];
-
-  const bulkActionsMenuItems: MenuProps['items'] = [
-    {
-      icon: <CopyClipboardIcon />,
-      key: 'copy',
-      label: t('Copy')
-    },
-    {
-      icon: <DataMovementIcon />,
-      key: 'move',
-      label: t('Move')
-    },
-    {
-      icon: <DownloadIcon />,
-      key: 'download',
-      label: t('Download')
-    },
-    {
-      icon: <DeleteIcon />,
-      key: 'delete',
-      label: t('Delete')
-    }
-  ];
-
-  const handleCreateNewFolder = (folderName: string) => {
-    setLoadingFiles(true);
-    mkdir(folderName, filePath)
-      .then(() => {
-        setRefreshKey(oldKey => oldKey + 1);
-      })
-      .catch(error => {
-        // eslint-disable-next-line no-restricted-syntax
-        console.log(error);
-      })
-      .finally(() => {
-        setLoadingFiles(false);
-      });
-  };
-
-  const handleCreateNewFile = (fileName: string) => {
-    setLoadingFiles(true);
-    touch(fileName, filePath)
-      .then(() => {
-        setRefreshKey(oldKey => oldKey + 1);
-      })
-      .catch(error => {
-        // eslint-disable-next-line no-restricted-syntax
-        console.log(error);
-      })
-      .finally(() => {
-        setLoadingFiles(false);
-      });
-  };
 
   useEffect(() => {
     setLoadingFiles(true);
@@ -215,37 +110,6 @@ const StorageBrowserTabContent = ({
             showIcon={false}
           />
         </div>
-        <div className="hue-storage-browser__actions-bar">
-          <Input className="hue-storage-browser__search" placeholder={t('Search')} />
-          <div className="hue-storage-browser__actions-bar-right">
-            <Dropdown
-              overlayClassName="hue-storage-browser__actions-dropdown"
-              menu={{
-                items: bulkActionsMenuItems,
-                className: 'hue-storage-browser__action-menu'
-              }}
-              trigger={['hover', 'click']}
-            >
-              <Button data-event={''}>
-                {t('Bulk Actions')}
-                <DropDownIcon />
-              </Button>
-            </Dropdown>
-            <Dropdown
-              overlayClassName="hue-storage-browser__actions-dropdown"
-              menu={{
-                items: newActionsMenuItems,
-                className: 'hue-storage-browser__action-menu'
-              }}
-              trigger={['hover', 'click']}
-            >
-              <PrimaryButton data-event={''}>
-                {t('New')}
-                <DropDownIcon />
-              </PrimaryButton>
-            </Dropdown>
-          </div>
-        </div>
         <StorageBrowserTable
           dataSource={files}
           pageStats={pageStats}
@@ -257,22 +121,9 @@ const StorageBrowserTabContent = ({
           onSortOrderChange={setSortOrder}
           sortByColumn={sortByColumn}
           sortOrder={sortOrder}
-        />
-        <InputModal
-          title={t('Create New Folder')}
-          inputLabel={t('Enter Folder name here')}
-          submitText={t('Create')}
-          showModal={showNewFolderModal}
-          onSubmit={handleCreateNewFolder}
-          onClose={() => setShowNewFolderModal(false)}
-        />
-        <InputModal
-          title={t('Create New File')}
-          inputLabel={t('Enter File name here')}
-          submitText={t('Create')}
-          showModal={showNewFileModal}
-          onSubmit={handleCreateNewFile}
-          onClose={() => setShowNewFileModal(false)}
+          setRefreshKey={setRefreshKey}
+          setLoadingFiles={setLoadingFiles}
+          filePath={filePath}
         />
       </div>
     </Spin>
